+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.7.1]
+
+### Added
+
+- Summaries for intermediate ASE optimization steps are now stored in the output schema for VASP and cclib-based recipes when `store_intermediate_results=True`
+- Added a `WRITE_PICKLE` setting, which by default, will write out a `.pkl` file containing the result schema to the directory
+
 ## [0.7.0]
 
 ### Added